added support for station stopletters for map & tableview

+++ tflapp/ViewController/TFLNearbyBusStationsController/TFLBusPredictionView.swift
@@ -20,7 +20,7 @@ class TFLBusPredictionView: UICollectionView {
     
     func setPredictions( predictions : [TFLBusStopArrivalsViewModel.LinePredictionViewModel], animated: Bool = false) {
         let visiblePredictions = Array(predictions[0..<min(predictions.count,self.maxVisibleCells)])
-        if  !animated || self.predictions.isEmpty {
+        if  !animated || self.predictions.isEmpty || self.predictions ==  visiblePredictions {
             self.predictions = visiblePredictions
             self.reloadData()
         }
@@ -28,28 +28,34 @@ class TFLBusPredictionView: UICollectionView {
             Crashlytics.log("oldTuples:\(self.predictions.map { $0.identifier }.joined(separator: ","))\nnewTuples:\(visiblePredictions.map { $0.identifier }.joined(separator: ","))")
             
             let (inserted ,deleted ,updated, moved) = self.evaluateLists(oldList: self.predictions, newList: visiblePredictions, compare : TFLBusStopArrivalsViewModel.LinePredictionViewModel.compare)
-            self.performBatchUpdates({ [weak self] in
-                Crashlytics.notify()
-                self?.predictions = visiblePredictions
-                let insertedIndexPaths = inserted.map { IndexPath(item: $0.index,section:0)}
-                self?.insertItems(at: insertedIndexPaths )
-                moved.forEach { self?.moveItem(at: IndexPath(item: $0.oldIndex,section:0), to:  IndexPath(item: $0.newIndex,section:0)) }
-                let deletedIndexPaths = deleted.map { IndexPath(item: $0.index,section:0)}
-                self?.deleteItems(at: deletedIndexPaths)
-                } ,completion: { [weak self]  _ in
-                    self?.performBatchUpdates({ [weak self] in
-                        Crashlytics.notify()
-                        let updatedIndexPaths = updated.map { IndexPath(item: $0.index,section:0)}
-                        let movedIndexPaths = moved.map { IndexPath(item: $0.newIndex,section:0)}
-                        (updatedIndexPaths+movedIndexPaths).forEach { indexPath in
-                            let cell = self?.cellForItem(at: indexPath)
-                            self?.configure(cell, at: indexPath)
-                        }
-                        },completion: { _ in
+            if inserted.isEmpty && moved.isEmpty && deleted.isEmpty {
+                self.reloadData()
+            }
+            else {
+                self.performBatchUpdates({ [weak self] in
+                    Crashlytics.notify()
+                    self?.predictions = visiblePredictions
+                    let insertedIndexPaths = inserted.map { IndexPath(item: $0.index,section:0)}
+                    self?.insertItems(at: insertedIndexPaths )
+                    moved.forEach { self?.moveItem(at: IndexPath(item: $0.oldIndex,section:0), to:  IndexPath(item: $0.newIndex,section:0)) }
+                    let deletedIndexPaths = deleted.map { IndexPath(item: $0.index,section:0)}
+                    self?.deleteItems(at: deletedIndexPaths)
+                    } ,completion: { [weak self]  _ in
+                        self?.performBatchUpdates({ [weak self] in
                             Crashlytics.notify()
-                    })
-            })
-       }
+                            let updatedIndexPaths = updated.map { IndexPath(item: $0.index,section:0)}
+                            let movedIndexPaths = moved.map { IndexPath(item: $0.newIndex,section:0)}
+                            (updatedIndexPaths+movedIndexPaths).forEach { indexPath in
+                                let cell = self?.cellForItem(at: indexPath)
+                                self?.configure(cell, at: indexPath)
+                            }
+                            },completion: { _ in
+                                Crashlytics.notify()
+                        })
+                })
+                
+            }
+        }
     }
     public var predictions : [TFLBusStopArrivalsViewModel.LinePredictionViewModel] = [] 
 }
